% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Bm}
\alias{Bm}
\title{Calculate between-imputation variance of the response mean
\deqn{\frac{\sum (\hat{Q}_l-\bar{Q})}{m-1}}}
\usage{
Bm(qhats, m)
}
\arguments{
\item{qhats}{vector of Qhats(means of response for each imputed dataset)}

\item{m}{number of imputed datasets}
}
\value{
Bm: the between-dataset variance of the response mean
}
\description{
Calculate between-imputation variance of the response mean
\deqn{\frac{\sum (\hat{Q}_l-\bar{Q})}{m-1}}
}
\examples{
imp = mice::mice(mice::nhanes)
qhats = Qhats(imp, "hyp")
m = imp$m
Bm(qhats, m)

}
