% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Qbar}
\alias{Qbar}
\title{Calculate Qbar (average response over MICE datasets)}
\usage{
Qbar(qhats)
}
\arguments{
\item{qhats}{vector of Qhats(response means for each imputed dataset)}
}
\value{
Qbar: the average response over MICEd datasets.
}
\description{
Calculate Qbar (average response over MICE datasets)
}
\examples{
imp = mice::mice(mice::nhanes)
qhats = Qhats(imp, "hyp")
Qbar(qhats)

}
