% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Qhats}
\alias{Qhats}
\title{Calculate Qhats (means of response for each imputed dataset)}
\usage{
Qhats(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of binary response variable}
}
\value{
Qhats: vector of response means for each imputed dataset
}
\description{
Calculate Qhats (means of response for each imputed dataset)
}
\examples{
imp = mice::mice(mice::nhanes)
Qhats(imp, "hyp")

}
