% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Rm}
\alias{Rm}
\title{Helper function for getting rm, a key component for
calculating degrees of freedom and the wilson CI directly}
\usage{
Rm(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of response variable}
}
\value{
rm
}
\description{
Helper function for getting rm, a key component for
calculating degrees of freedom and the wilson CI directly
}
\examples{
imp = mice::mice(mice::nhanes)
Rm(imp, "hyp")

}
