% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Ubar}
\alias{Ubar}
\title{Calculate Ubar (average response variance over MICE datasets)}
\usage{
Ubar(qhats, m, nrow)
}
\arguments{
\item{qhats}{vector of Qhats(means of response for each imputed dataset)}

\item{m}{number of imputed datasets}

\item{nrow}{number of observations in the imputed dataset}
}
\value{
Ubar: average response variance over MICE datasets
}
\description{
Calculate Ubar (average response variance over MICE datasets)
}
\examples{
imp = mice::mice(mice::nhanes)
qhats = Qhats(imp, "hyp")
m = imp$m
nrow = imp$data \%>\% nrow()
Ubar(qhats, m, nrow)

}
