% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Uhats}
\alias{Uhats}
\title{Calculate Uhats (variance for each imputed dataset)}
\usage{
Uhats(qhats, nrow)
}
\arguments{
\item{qhats}{vector of Qhats(means of response for each imputed dataset)}

\item{nrow}{number of observations in the imputed dataset}
}
\value{
Uhats: vector of response variances for each imputed dataset
}
\description{
Calculate Uhats (variance for each imputed dataset)
}
\examples{
imp = mice::mice(mice::nhanes)
qhats = Qhats(imp, "hyp")
nrow = imp$data \%>\% nrow()
Uhats(qhats, nrow)

}
