% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Bm}
\alias{Bm}
\title{Calculate between-imputation variance of the response mean
\deqn{\frac{\sum (\hat{Q}_l-\bar{Q})}{m-1}}}
\usage{
Bm(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of response variable}
}
\value{
Bm: the between-dataset variance of the response mean
}
\description{
Calculate between-imputation variance of the response mean
\deqn{\frac{\sum (\hat{Q}_l-\bar{Q})}{m-1}}
}
\examples{
imp = mice::mice(mice::nhanes)
Bm(imp, "bmi")

}
