% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{dof}
\alias{dof}
\title{Calculate degrees of freedom used in calculating
confidence intervals of t-distributed proportion point
estimate \eqn{\bar{Q}_m}}
\usage{
dof(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of response variable}
}
\value{
degrees of freedom
}
\description{
Calculate degrees of freedom used in calculating
confidence intervals of t-distributed proportion point
estimate \eqn{\bar{Q}_m}
}
\examples{
imp = mice::mice(mice::nhanes)
dof(imp, "hyp")

}
