% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{mi_wald}
\alias{mi_wald}
\title{Calculates the specified Wald CI of a binomial proportion
variable, given imputed data sets.}
\usage{
mi_wald(mids_obj, response, ci_level)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of response variable}

\item{ci_level}{desired confidence interval level}
}
\value{
two-length vector of lower CI and upper CI
}
\description{
Calculates the specified Wald CI of a binomial proportion
variable, given imputed data sets.
}
\examples{
imp = mice::mice(mice::nhanes)
mi_wald(imp, "hyp", 0.95)

}
