% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Qbar}
\alias{Qbar}
\title{Calculate Qbar (average response over MICE datasets)}
\usage{
Qbar(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of endogenous variable}
}
\value{
Qbar: the average response over MICEd datasets.
}
\description{
Calculate Qbar (average response over MICE datasets)
}
\examples{
imp = mice::mice(mice::nhanes)
Qbar(imp, "hyp")

}
