% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Rm}
\alias{Rm}
\title{Helper function for getting rm, a key component for
calculating degrees of freedom and the wilson CI directly}
\usage{
Rm(qhats, m, nrow)
}
\arguments{
\item{qhats}{vector of Qhats(means of response for each imputed dataset)}

\item{m}{number of imputed datasets}

\item{nrow}{number of observations in the imputed dataset}
}
\value{
rm
}
\description{
Helper function for getting rm, a key component for
calculating degrees of freedom and the wilson CI directly
}
\examples{
imp = mice::mice(mice::nhanes)
qhats = Qhats(imp, "hyp")
m = imp$m
nrow = imp$data \%>\% nrow()
Rm(qhats, m, nrow)

}
