% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Tm}
\alias{Tm}
\title{Estimate variance of proportion point estimate \eqn{\bar{Q}_m}}
\usage{
Tm(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of response variable}
}
\value{
variance of proportion point estimate
}
\description{
Estimate variance of proportion point estimate \eqn{\bar{Q}_m}
}
\examples{
imp = mice::mice(mice::nhanes)
Tm(imp, "bmi")

}
