% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Ubar}
\alias{Ubar}
\title{Calculate Ubar (average response variance over MICE datasets)}
\usage{
Ubar(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of endogenous variable}
}
\value{
Ubar: average response variance over MICE datasets
}
\description{
Calculate Ubar (average response variance over MICE datasets)
}
\examples{
imp = mice::mice(mice::nhanes)
Ubar(imp, "bmi")
}
