% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilson.R
\name{Uhats}
\alias{Uhats}
\title{Calculate Uhats (variance for each imputed dataset)}
\usage{
Uhats(mids_obj, response)
}
\arguments{
\item{mids_obj}{mids object created by mice package}

\item{response}{string name of binary response variable}
}
\value{
Uhats: vector of response variances for each imputed dataset
}
\description{
Calculate Uhats (variance for each imputed dataset)
}
\examples{
imp = mice::mice(mice::nhanes)
Uhats(imp,"hyp")

}
